Fix concurrent_writes_are_serialized
Before, it was possible for the listener
to accept the connection from p1 instead of p0.
If that happened, the test would get stuck when
reading messages.
Now, we accept connections from both p0 and p1
and check the messages in two tasks concurrently.
One of these tasks will receive no messages (from p1),
the other (if correct) will break out of the loop and
trigger `receiving_set.join_next().await`